+++ SDPC/src/main/java/skyclash/skyclash/fileIO/MOTD.java
@@ -11,7 +11,6 @@ import java.util.Random;
 
 import org.bukkit.Bukkit;
 
-import net.md_5.bungee.api.ChatColor;
 import skyclash.skyclash.Scheduler;
 
 public class MOTD {
@@ -21,7 +20,6 @@ public class MOTD {
         if (!(new File(motdBankPath).exists())) {
             return;
         }
-        Bukkit.getConsoleSender().sendMessage(ChatColor.GRAY+"MOTD Changed");
 
         ArrayList<String> motds = readFile(motdBankPath);
         String motd = motds.get(new Random().nextInt(motds.size()));